Revert "Don't use `raw` in comparisons."
This reverts commit 9ff9176b6707c259c2546c2500867181372ce6da.

Per the discussion in #6, having the `Eq` and `Ord` instances ignore the `raw`
field of `Ident` causes more trouble than it's worth, as it causes the parser
to incorrectly deem raw identifiers like `r#return` to be keywords. While we
could fix this issue by changing the parser, this would take quite a bit of
code changes to accomplish. As such, we revert the change here, and we make a
note in the Haddocks for the `Eq` and `Ord` instances to beware of the fact
that `raw` is taken into account.

After this change, the `rustc-tests` test suite passes once more. As such, this
change fixes #6.